extracted magick code
